+++ useintest/predefined/consul/__init__.py
@@ -1,2 +1,2 @@
 from useintest.predefined.consul.consul import ConsulServiceController, consul_service_controllers, \
-    Consul1_0_0ServiceController, Consul0_8_4ServiceController
+    Consul1_0_0ServiceController, Consul0_8_4ServiceController, ConsulDockerisedService