+++ useintest/tests/predefined/consul/test_consul.py
@@ -1,10 +1,10 @@
 import unittest
 from abc import ABCMeta
 
-from consul import Consul
 from hgicommon.testing import TypeUsedInTest, create_tests, get_classes_to_test
 
-from useintest.predefined.consul.consul import ConsulServiceController, consul_service_controllers
+from useintest.predefined.consul.consul import ConsulServiceController, consul_service_controllers, \
+    ConsulDockerisedService
 from useintest.services.models import DockerisedServiceWithUsers
 from useintest.tests.services.common import TestServiceControllerSubclass
 
@@ -13,13 +13,14 @@ _TEST_VALUE = "world"
 
 
 class _TestConsulServiceController(
-    TestServiceControllerSubclass[TypeUsedInTest, DockerisedServiceWithUsers], metaclass=ABCMeta):
+        TestServiceControllerSubclass[TypeUsedInTest, DockerisedServiceWithUsers], metaclass=ABCMeta):
     """
     Tests for `ConsulServiceController`.
     """
     def test_start(self):
-        service = self._start_service()
-        consul_client = Consul(service.host, service.ports[8500])
+        service = self._start_service()     # type: ConsulDockerisedService
+        self.assertIsInstance(service, ConsulDockerisedService)
+        consul_client = service.create_consul_client()
         consul_client.kv.put(_TEST_KEY, _TEST_VALUE)
         self.assertEqual(_TEST_VALUE, consul_client.kv.get(_TEST_KEY)[1]["Value"].decode("utf-8") )